Removing implicit casts.
Whether or not these were a problem to begin with is Juan's problem.

+++ Src/Material/Material.cpp
@@ -23,7 +23,7 @@ Texture* Material::getTexture(int index) const {
 }
 
 int Material::getTextureCount() const {
-    return textures.size();
+    return (int)textures.size();
 }
 
 bool Material::isOpaque() const {
